+++ controllers/create_address.go
@@ -2,7 +2,6 @@ package controllers
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 
 	"bitbucket.org/andyfusniakteam/ecom-api-go/services"
@@ -10,14 +9,14 @@ import (
 )
 
 type addressRequestBody struct {
-	Typ         string `json:"typ"`
-	ContactName string `json:"contact_name"`
-	Addr1       string `json:"address_line_1"`
-	Addr2       string `json:"address_line_2"`
-	City        string `json:"city"`
-	County      string `json:"county"`
-	Postcode    string `json:"postcode"`
-	Country     string `json:"country"`
+	Typ         string  `json:"typ"`
+	ContactName string  `json:"contact_name"`
+	Addr1       string  `json:"addr1"`
+	Addr2       *string `json:"addr2"`
+	City        string  `json:"city"`
+	County      *string `json:"county"`
+	Postcode    string  `json:"postcode"`
+	Country     string  `json:"country"`
 }
 
 // CreateAddress handler
@@ -29,16 +28,12 @@ func CreateAddress(w http.ResponseWriter, r *http.Request) {
 
 	params := mux.Vars(r)
 
-	fmt.Println(params["cid"])
-
 	o := addressRequestBody{}
 	err := json.NewDecoder(r.Body).Decode(&o)
 	if err != nil {
 		http.Error(w, err.Error(), 400)
 		return
 	}
-
-	fmt.Println(o)
 
 	address := services.CreateAddress(params["cid"], o.Typ, o.ContactName, o.Addr1, o.Addr2, o.City, o.County, o.Postcode, o.Country)