+++ controllers/get_customer.go
@@ -2,7 +2,6 @@ package controllers
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 
 	"bitbucket.org/andyfusniakteam/ecom-api-go/services"
@@ -12,9 +11,6 @@ import (
 // GetCustomer controller
 func GetCustomer(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
-
-	fmt.Println(r.Host)
-	fmt.Println(r.URL)
 
 	customer := services.GetCustomer(params["cid"])