+++ scripts/ecom_test_env.sh
@@ -5,3 +5,5 @@ export ECOM_DBUSER=postgres
 export ECOM_DBPASS=postgres
 export ECOM_DBNAME=ecom_dev
 export ECOM_SSL=disable
+export ECOM_CREDENTIALS_FILE=/home/andy/ecom-test-fa3e406ce4fe.json
+