+++ services/customers.go
@@ -2,11 +2,35 @@ package services
 
 import (
 	"bitbucket.org/andyfusniakteam/ecom-api-go/models"
+	firebase "firebase.google.com/go"
+	"firebase.google.com/go/auth"
+	"golang.org/x/net/context"
 )
 
+// App is a Firebase App used for handling authentication
+var App *firebase.App
+
 // CreateCustomer creates a new customer
-func CreateCustomer(firstname, lastname string) (*models.Customer, error) {
-	return models.CreateCustomer(firstname, lastname)
+func CreateCustomer(email, password, firstname, lastname string) (*models.Customer, error) {
+	ctx := context.Background()
+	authClient, err := App.Auth(ctx)
+	if err != nil {
+		panic(err)
+	}
+
+	user := (&auth.UserToCreate{}).
+		Email(email).
+		EmailVerified(false).
+		Password(password).
+		DisplayName(`${firstname} ${lastname}`).
+		Disabled(false)
+
+	userRecord, err := authClient.CreateUser(ctx, user)
+	if err != nil {
+		panic(err)
+	}
+
+	return models.CreateCustomer(userRecord.UID, email, firstname, lastname)
 }
 
 // GetCustomer retrieves a customer by customer UUID
@@ -15,10 +39,8 @@ func GetCustomer(customerUUID string) *models.Customer {
 }
 
 // CreateAddress creates a new address for a customer
-func CreateAddress(customerUUID string, typ string, contactName string, addr1 string,
-	addr2 string, city string, county string, postcode string, country string) *models.Address {
+func CreateAddress(customerUUID, typ, contactName, addr1 string, addr2 *string, city string, county *string, postcode string, country string) *models.Address {
 	customerID, _ := models.GetCustomerIDByUUID(customerUUID)
-
 	return models.CreateAddress(customerID, typ, contactName, addr1, addr2, city, county, postcode, country)
 }