+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.3.0 (7 Nov 2018)
++ Firebase Auth Go library calls CreateUser
++ Google Stackdriver logging
++ Address addr2 and county fields are optional
+
 ## 0.2.0 (1 Nov 2018)
 + GetCartItems operation returns all items in a a shopping cart
 + AddItemToCart uses the products_pricing table to lookup the price using the default tier